// Online_Pharmacy_and_Delivery_Frontend/src/shared/types/api/admin.d.ts
export interface components {
    schemas: {
// ...
        OrderResponseDto: {
            /** Format: int64 */
            id?: number;
            /** Format: int64 */
            userId?: number;
            userEmail?: string;
            userName?: string;
            status?: string;
            /** Format: double */
            totalAmount?: number;
            /** Format: double */
            discountAmount?: number;
            /** Format: double */
            taxAmount?: number;
            /** Format: double */
            deliveryCharge?: number;
            items?: components["schemas"]["OrderItemResponseDto"][];
            deliveryAddress?: string;
            pincode?: string;
            deliverySlot?: string;
            /** Format: int64 */
            prescriptionId?: number;
            adminNote?: string;
            paymentMethod?: string;
            paymentId?: string;
            createdAt?: string;
            updatedAt?: string;
            deliveredAt?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}